# specs/contracts/api.yaml
openapi: "3.1.0"
info:
  title: smartLattes API
  version: "2.0.0"
  description: API for uploading Lattes XML CVs, generating AI-powered researcher summaries, and exporting documents
servers:
  - url: http://localhost:8080
    description: Local development / Docker container
paths:
...
  /api/summary/save:
    post:
      summary: Save generated summary
      description: |
        Stores the generated summary in the resumos MongoDB collection.
        Called after user confirms (Ok or download).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveSummaryRequest"
      responses:
        "200":
          description: Summary saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveSummaryResponse"
        "400":
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: MongoDB unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    SaveSummaryRequest:
      type: object
      required:
        - lattesId
        - summary
        - provider
        - model
      properties:
        lattesId:
          type: string
        summary:
          type: string
        provider:
          type: string
        model:
          type: string
    SaveSummaryResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Resumo salvo com sucesso"
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
...